[Update] : fixed select to render page when option changes

+++ src/views/modifyCategory/modifyCategory.js
@@ -5,14 +5,12 @@ const selectBox = document.querySelector("#selectBox");
 const submitBtn = document.querySelector("#btnSubmit");
 
 // Data mocked
-const {accessToken, userId, userType} = {
-  "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2MzY4YmRmNzQ5ZGZlODA4OTg4ZWEyMTIiLCJyb2xlIjoiYWRtaW4iLCJpYXQiOjE2Njc4NzU1MTh9.jaHZAFtGgXdvUWw_QoQyWztNjIQGCCyOZZd_mWlTbuU",
-  "userId": "6368bdf749dfe808988ea212",
-  "userType": "admin"
-};
-sessionStorage.setItem('token', accessToken)
-
-var categories = await getCategory();
+// const {accessToken, userId, userType} = {
+//   "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2MzY4YmRmNzQ5ZGZlODA4OTg4ZWEyMTIiLCJyb2xlIjoiYWRtaW4iLCJpYXQiOjE2Njc4OTk2NTYsImV4cCI6MTY2Nzk4NjA1Nn0.A62VUIXHdHdJocM-ws40oEeq4ppj9N3sf6tfP33WWmI",
+//   "userId": "6368bdf749dfe808988ea212",
+//   "userType": "admin"
+// }
+// sessionStorage.setItem('token', accessToken)
 // const categories = [`
 // <option>1</option>
 // <option>2</option>
@@ -24,11 +22,7 @@ var categories = await getCategory();
 // `, 7, [1,2,3,4,5,6,7]]
 
 
-
-
-// const {accessToken, userId, userType} = JSON.parse(sessionStorage.getItem('loginToken'));
-// const emailToken = sessionStorage.getItem('userEmail')
-
+var categories = await getCategory();
 addAllElements();
 addAllEvents();
 
@@ -41,23 +35,22 @@ function addAllElements() {
 function addAllEvents() {
   catchselect.addEventListener('change', categoryModify)
   submitBtn.addEventListener('click', submitModify)
-  submitBtn.addEventListener('click', updateCategory)
 }
-
-async function updateCategory () {
-  categories = await getCategory();
-}
-
 
 
 async function categoryModify () {
-  const numCategory = categories[1];
-  if (numCategory > 7 & catchselect.value === "생성")
-    return alert("카테고리는 8개 이하만 생성해 주세요.")
-  else
-  {
-    selectBox.innerHTML = template(catchselect.value);  
+  if (categories[1] > 7 & catchselect.value === "생성"){
+    catchselect.value = "수정"
+    alert("카테고리는 8개 이하만 생성해 주세요.")
   }
+  else if (categories[1] === 0 & catchselect.value === "삭제"){
+    catchselect.value = "수정"
+    alert("삭제할 카테고리가 없습니다.")
+  }
+  let updateCategory = await getCategory();
+  console.log(updateCategory);
+  selectBox.innerHTML = template(catchselect.value, updateCategory[0]);
+  return updateCategory;
 }
 
 async function submitModify () {
@@ -68,9 +61,9 @@ async function submitModify () {
   {
     if (categories[1] > 7){
       alert ("카테고리를 더 추가할 수 없습니다.")
-      catchselect.value = "생성"
       return
     }
+
     const name = document.querySelector('.input').value;
     if (name.length > 6)
       return alert ("카테고리 이름은 여섯 글자 이하입니다.")
@@ -78,46 +71,62 @@ async function submitModify () {
     const data = {
       name,
     }
+
     const res = await Api.post('/admin/categories', data)
-    categories = await getCategory();
     if (res.statusCode)
       return alert(res.reason);
-    else
+    else{
+      categories = await categoryModify();
       return alert("성공적으로 등록되었습니다.")
+    }
   }
   // put
   else if (selectedAction === 1)
   {
     const categoryInd = document.querySelectorAll('select')[1].selectedIndex;
+    const name = document.querySelector('.input').value;
+    if (name.length > 6)
+      return alert ("카테고리 이름은 여섯 글자 이하입니다.")
+    
     const data = {
-      name : document.querySelector('.input').value,
+      name,
     }
+
     const catId = categories[2][categoryInd];
     const res = await Api.put(`/admin/categories/${catId}`, data)
-    categories = await getCategory();
-    console.log("modify response : ",res)
+
     if (res.statusCode)
       return alert(res.reason);
-    else
+    else{
+      categories = await categoryModify();
       return alert("성공적으로 수정되었습니다.")
+    }
   }
   // delete
   else if (selectedAction === 2)
   {
+    if (categories[1] < 1){
+      alert ("카테고리를 더 삭제할 수 없습니다.")
+      return
+    }
+
     const categoryInd = document.querySelectorAll('select')[1].selectedIndex;
     const catId = categories[2][categoryInd];
     const res = await Api.delete(`/admin/categories/${catId}`)
-    categories = await getCategory();
-    console.log("delete Response : ",res);
-    if (res.success)
+
+    if (res.success){
+      categories = await categoryModify();
       return alert("성공적으로 삭제되었습니다.");
+    }
     else
       return alert(res.reason)
   }
 }
 
 
-function template(option) {
+// Below are helper functions
+
+function template(option, categoryInfo) {
   if (option === "생성")
     return `<div class="field">
     <label class="label">카테고리 생성</label>
@@ -131,7 +140,7 @@ function template(option) {
     <div class="control">
       <div class="select is-primary">
         <select>
-          ${categories}
+          ${categoryInfo}
         </select>
       </div>
     </div>
@@ -148,7 +157,7 @@ function template(option) {
     <div class="control">
       <div class="select is-primary">
         <select>
-          ${categories}
+          ${categoryInfo}
         </select>
       </div>
     </div>
@@ -182,12 +191,12 @@ function initializeForm(numCategory)
   if (numCategory > 7)
   {
     catchselect.value = "수정"
-    selectBox.innerHTML = template("수정")
+    selectBox.innerHTML = template("수정", categories[0])
   }
   else
   {
     catchselect.value = "생성"
-    selectBox.innerHTML = template("생성")
+    selectBox.innerHTML = template("생성", categories[0])
   }
 }
